backend/codeXpose/requirements.txt
Django==2.0
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.10
mysqlclient==1.3.12
pkg-resources==0.0.0
pytz==2017.3
